test: isolate filesystem and codec coverage
Keep unit timing deterministic by running filesystem-backed plan execution and real image-decoding fallback coverage in the integration lane.
