+++ test/utils/image/backend/resolveImageBackend.test.ts
@@ -33,28 +33,6 @@ describe("resolveImageBackend", () => {
     expect(resolveImageBackend({ platform: "linux" })).toBeInstanceOf(SharpBackend);
   });
 
-  test("falls back to JimpBackend behavior when sharp discovery fails", async () => {
-    const { Jimp, rgbaToInt } = await import("jimp");
-    const source = await new Jimp({
-      width: 2,
-      height: 2,
-      color: rgbaToInt(1, 2, 3, 255),
-    }).getBuffer("image/png");
-    const backend = resolveImageBackend({
-      platform: "linux",
-      sharpLoader: async () => {
-        throw new Error("sharp unavailable");
-      },
-    });
-
-    expect(backend).toBeInstanceOf(SharpBackend);
-    await expect(backend.metadata(source)).resolves.toMatchObject({
-      width: 2,
-      height: 2,
-      format: "png",
-    });
-  });
-
   test("returns a fresh instance per call", () => {
     expect(resolveImageBackend()).not.toBe(resolveImageBackend());
   });
